MINOR CLIENTS: The Facade In Cohesive Partials - Verbs, Runs, Composition
