+++ Standard.Agents/StandardAgent.cs
@@ -1127,169 +1127,6 @@ public sealed partial class StandardAgent : IAgent
         Set(() => this.loggingBroker = broker);
 
     /// <summary>
-    /// Runs the agent on a prompt to completion and returns the final answer. The first call
-    /// composes the configured pieces (brain, skills, tools, guardians, memory, knowledge); later
-    /// calls reuse that composition unless a builder method changed the configuration.
-    /// </summary>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <returns>The agent's final answer.</returns>
-    // async, so a composition failure surfaces on await rather than being thrown
-    // synchronously out of a method whose signature promises a ValueTask. A caller
-    // doing `var task = agent.ProcessPromptAsync(p); ... await task;` would otherwise
-    // be hit at the assignment, nowhere near the await they were guarding.
-    public async ValueTask<string> ProcessPromptAsync(string prompt)
-    {
-        return await (await ResolveAgentAsync()).ProcessPromptAsync(prompt);
-    }
-
-    /// <summary>
-    /// Runs the agent and reports <b>how the run ended</b> as well as what it produced. Only
-    /// <c>AgentStatus.Responded</c> makes the result an answer — a run held on an authority,
-    /// refused, cancelled or out of turns produced prose about why, and the two read alike.
-    /// </summary>
-    /// <remarks>
-    /// This is what nesting uses. <c>AgentTool</c> reads it so an outer agent cannot mistake a
-    /// sub-agent that was held for one that finished, which the string on its own could never tell
-    /// it.
-    /// </remarks>
-    /// <param name="prompt">What to work on.</param>
-    /// <returns>The answer, and how the run ended.</returns>
-    public async ValueTask<AgentOutcome> RunAsync(string prompt) =>
-        await (await ResolveAgentAsync()).RunAsync(prompt, string.Empty, CancellationToken.None);
-
-    /// <summary>
-    /// The same run, stoppable: cancellation stops it at the next turn boundary, so no effect
-    /// is left half-recorded (SPEC.md §4.10). This is the overload nesting calls —
-    /// <c>AgentTool</c> forwards the outer run's token here, so cancelling an outer run stops
-    /// the whole tree.
-    /// </summary>
-    /// <param name="prompt">What to work on.</param>
-    /// <param name="cancellationToken">Token to stop the run.</param>
-    /// <returns>The answer, and how the run ended.</returns>
-    public async ValueTask<AgentOutcome> RunAsync(
-        string prompt,
-        CancellationToken cancellationToken) =>
-        await (await ResolveAgentAsync()).RunAsync(prompt, string.Empty, cancellationToken);
-
-    /// <summary>
-    /// Runs a caller's request and reports <b>how the run ended</b> as well as what it produced.
-    /// This is the exposer's read: a run that ended <c>AwaitingInput</c> holding a caller's tool
-    /// call and a run that answered read alike as strings, and only the status tells them apart.
-    /// </summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <returns>The answer, and how the run ended.</returns>
-    public async ValueTask<AgentOutcome> RunAsync(PromptRequest request) =>
-        await RunAsync(request, CancellationToken.None);
-
-    /// <summary>The same request-carrying run, reported with how it ended, cancellable.</summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <param name="cancellationToken">Token to stop the run.</param>
-    /// <returns>The answer, and how the run ended.</returns>
-    public async ValueTask<AgentOutcome> RunAsync(
-        PromptRequest request,
-        CancellationToken cancellationToken) =>
-        await (await ResolveAgentAsync()).RunAsync(request, cancellationToken);
-
-    /// <summary>
-    /// Runs the agent on a prompt and stops when <paramref name="cancellationToken"/> is
-    /// cancelled — at the next turn boundary at the latest, so no effect is left half-recorded
-    /// (SPEC.md §4.10). A cancelled run returns a message saying so rather than an answer.
-    /// </summary>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <param name="cancellationToken">Token to stop the run.</param>
-    /// <returns>The agent's final answer, or a message explaining why it stopped.</returns>
-    public async ValueTask<string> ProcessPromptAsync(
-        string prompt,
-        CancellationToken cancellationToken)
-    {
-        return await (await ResolveAgentAsync()).ProcessPromptAsync(prompt, cancellationToken);
-    }
-
-    /// <summary>
-    /// Runs the agent as part of a <b>conversation</b>. What was said before is loaded before the
-    /// brain thinks, and this exchange is appended when it answers — so a follow-up resolves
-    /// against what came before instead of starting from nothing (SPEC.md §4.11).
-    /// </summary>
-    /// <remarks>
-    /// The conversation lives in the session store, not in this instance, so it survives a
-    /// restart and can be continued by a different process. Invariant 4 is intact: the agent is
-    /// still stateless across prompts; the session is not part of the agent.
-    /// <para>A cancelled or budget-stopped run is never recorded as an answer — the next prompt
-    /// would otherwise be told the agent said something it never said.</para>
-    /// </remarks>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <param name="sessionId">Which conversation this belongs to.</param>
-    /// <param name="cancellationToken">Token to stop the run.</param>
-    /// <returns>The agent's final answer.</returns>
-    public async ValueTask<string> ProcessPromptAsync(
-        string prompt,
-        string sessionId,
-        CancellationToken cancellationToken = default)
-    {
-        return await (await ResolveAgentAsync())
-            .ProcessPromptAsync(prompt, sessionId, cancellationToken);
-    }
-
-    /// <summary>
-    /// Runs the agent on a caller's <b>request</b> — a prompt carrying its own inference options
-    /// (docs/per-request-inference.md). What is established and hard-configured takes precedence,
-    /// always: a request can shape a run only where the deployment expressed no opinion, and it
-    /// can never widen the boundary the deployment set.
-    /// </summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <returns>The agent's final answer.</returns>
-    public async ValueTask<string> ProcessPromptAsync(PromptRequest request)
-    {
-        return await ProcessPromptAsync(request, CancellationToken.None);
-    }
-
-    /// <summary>
-    /// The same request-carrying run, stopped when <paramref name="cancellationToken"/> is
-    /// cancelled — at the next turn boundary at the latest (SPEC.md §4.10).
-    /// </summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <param name="cancellationToken">Token to stop the run.</param>
-    /// <returns>The agent's final answer.</returns>
-    public async ValueTask<string> ProcessPromptAsync(
-        PromptRequest request,
-        CancellationToken cancellationToken)
-    {
-        return await (await ResolveAgentAsync()).ProcessPromptAsync(request, cancellationToken);
-    }
-
-    /// <summary>
-    /// Continues a session that stopped in the middle of something — waiting on a person, waiting
-    /// on an authority, or killed outright (SPEC.md §4.9, §4.11).
-    /// </summary>
-    /// <remarks>
-    /// This is the same call as <see cref="ProcessPromptAsync(string, string, CancellationToken)"/>
-    /// and reads the answer the same way; it exists because <i>resuming</i> is what the caller is
-    /// doing, and a verb that says so is worth more than one saved line of documentation.
-    /// <para>A session that never delivered an answer keeps the interrupted run's identity, so an
-    /// act it already performed is recognised as its own and replayed rather than performed twice.
-    /// That only holds across processes if the ledger is durable — see
-    /// <see cref="EffectLedger(string)"/>; the built-in one lives in memory and dies with the
-    /// instance.</para>
-    /// <para>Nothing is required of the caller beyond the answer. There is no separate resume
-    /// mode and no state to hand back: the session already holds everything, which is what makes
-    /// a different process able to pick it up.</para>
-    /// </remarks>
-    /// <param name="sessionId">The conversation to continue.</param>
-    /// <param name="answer">
-    /// What the agent was waiting for — a reply to its question, or an authority's decision.
-    /// </param>
-    /// <param name="cancellationToken">Token to stop the run.</param>
-    /// <returns>The agent's answer, now that it can finish.</returns>
-    public async ValueTask<string> ResumeAsync(
-        string sessionId,
-        string answer,
-        CancellationToken cancellationToken = default)
-    {
-        return await (await ResolveAgentAsync())
-            .ProcessPromptAsync(answer, sessionId, cancellationToken);
-    }
-
-    /// <summary>
     /// Keeps conversations in a folder, one file per session — the <b>Local</b> mode of sessions.
     /// </summary>
     /// <param name="path">Folder to keep sessions in.</param>
@@ -1501,626 +1338,4 @@ public sealed partial class StandardAgent : IAgent
     /// <returns>The same agent, so calls can be chained.</returns>
     public StandardAgent UseResilience(IResilienceBroker broker) =>
         Set(() => this.resilienceBroker = broker);
-
-    /// <summary>
-    /// Runs the agent on a prompt and streams its progress as it happens — status updates, the
-    /// brain's thinking, and response text arrive as <see cref="AgentStreamEvent"/> values rather
-    /// than waiting for the final answer. Use this to surface a live view of the agent's work.
-    /// </summary>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <param name="cancellationToken">Token to stop streaming early.</param>
-    /// <returns>An async stream of events describing the agent's run.</returns>
-    // async iterator, so a composition failure surfaces when the caller starts
-    // enumerating rather than at the call site — mirroring ProcessPromptAsync.
-    public async IAsyncEnumerable<AgentStreamEvent> StreamPromptAsync(
-        string prompt,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default)
-    {
-        IAsyncEnumerable<AgentStreamEvent> events =
-            (await ResolveAgentAsync()).ProcessPromptStreamAsync(prompt, cancellationToken);
-
-        await foreach (AgentStreamEvent streamEvent in events.WithCancellation(cancellationToken))
-        {
-            yield return streamEvent;
-        }
-    }
-
-    /// <summary>
-    /// Streams the agent's work as part of a <b>conversation</b> — the streamed equivalent of
-    /// <see cref="ProcessPromptAsync(string, string, CancellationToken)"/>.
-    /// </summary>
-    /// <remarks>
-    /// Every control the batched call enforces holds here too: budgets, cancellation, session
-    /// loading and the record of the exchange. A control a caller can step around by changing
-    /// method is not a control (SPEC.md §7.6).
-    /// </remarks>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <param name="sessionId">Which conversation this belongs to.</param>
-    /// <param name="cancellationToken">Token to stop streaming early.</param>
-    /// <returns>An async stream of events describing the agent's run.</returns>
-    public async IAsyncEnumerable<AgentStreamEvent> StreamPromptAsync(
-        string prompt,
-        string sessionId,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default)
-    {
-        IAsyncEnumerable<AgentStreamEvent> events = (await ResolveAgentAsync())
-            .ProcessPromptStreamAsync(prompt, sessionId, cancellationToken);
-
-        await foreach (AgentStreamEvent streamEvent in events.WithCancellation(cancellationToken))
-        {
-            yield return streamEvent;
-        }
-    }
-
-    /// <summary>
-    /// Streams the agent's work on a caller's <b>request</b> — the streamed equivalent of
-    /// <see cref="ProcessPromptAsync(PromptRequest, CancellationToken)"/>. The same precedence,
-    /// the same controls: a control a caller can step around by changing method is not a control
-    /// (SPEC.md §7.6).
-    /// </summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <returns>An async stream of events describing the agent's run.</returns>
-    public IAsyncEnumerable<AgentStreamEvent> StreamPromptAsync(PromptRequest request) =>
-        StreamPromptAsync(request, CancellationToken.None);
-
-    /// <summary>The same request-carrying stream, cancellable.</summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <param name="cancellationToken">Token to stop streaming early.</param>
-    /// <returns>An async stream of events describing the agent's run.</returns>
-    public async IAsyncEnumerable<AgentStreamEvent> StreamPromptAsync(
-        PromptRequest request,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
-    {
-        IAsyncEnumerable<AgentStreamEvent> events =
-            (await ResolveAgentAsync()).ProcessPromptStreamAsync(request, cancellationToken);
-
-        await foreach (AgentStreamEvent streamEvent in events.WithCancellation(cancellationToken))
-        {
-            yield return streamEvent;
-        }
-    }
-
-    /// <summary>
-    /// The streamed outcome (SPEC.md §4.14): the same run as <see cref="StreamPromptAsync(string, CancellationToken)"/>,
-    /// whose enumeration also carries — once it completes — the same structured outcome
-    /// <see cref="RunAsync(PromptRequest, CancellationToken)"/> returns: status, result, and any
-    /// pending effect with its model-minted call id. One run, two readings; a caller never
-    /// chooses between the answer's structure and the run's story.
-    /// </summary>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <param name="cancellationToken">Token to stop the run at its next turn boundary.</param>
-    /// <returns>The run's events, carrying its outcome at completion.</returns>
-    public AgentRunStream RunStreamAsync(
-        string prompt,
-        CancellationToken cancellationToken = default) =>
-        RunStreamAsync(new PromptRequest { Prompt = prompt }, cancellationToken);
-
-    /// <summary>The same streamed outcome, inside a conversation.</summary>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <param name="sessionId">Which conversation this belongs to.</param>
-    /// <returns>The run's events, carrying its outcome at completion.</returns>
-    public AgentRunStream RunStreamAsync(string prompt, string sessionId) =>
-        RunStreamAsync(prompt, sessionId, CancellationToken.None);
-
-    /// <summary>The same conversation-carrying streamed outcome, cancellable.</summary>
-    /// <param name="prompt">The user's prompt.</param>
-    /// <param name="sessionId">Which conversation this belongs to.</param>
-    /// <param name="cancellationToken">Token to stop the run at its next turn boundary.</param>
-    /// <returns>The run's events, carrying its outcome at completion.</returns>
-    public AgentRunStream RunStreamAsync(
-        string prompt,
-        string sessionId,
-        CancellationToken cancellationToken) =>
-        RunStreamAsync(
-            new PromptRequest { Prompt = prompt, SessionId = sessionId },
-            cancellationToken);
-
-    /// <summary>The same streamed outcome, asked for by a caller's request.</summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <returns>The run's events, carrying its outcome at completion.</returns>
-    public AgentRunStream RunStreamAsync(PromptRequest request) =>
-        RunStreamAsync(request, CancellationToken.None);
-
-    /// <summary>The same request-carrying streamed outcome, cancellable.</summary>
-    /// <param name="request">The caller's prompt and per-request inference options.</param>
-    /// <param name="cancellationToken">Token to stop the run at its next turn boundary.</param>
-    /// <returns>The run's events, carrying its outcome at completion.</returns>
-    public AgentRunStream RunStreamAsync(
-        PromptRequest request,
-        CancellationToken cancellationToken) =>
-        new(setOutcome => ResolvedRunStreamAsync(request, setOutcome, cancellationToken));
-
-    // The inner stream's outcome is handed outward when the enumeration ends, so the wrapper
-    // the caller holds carries what the composed agent's run concluded — same seam, one tier up.
-    private async IAsyncEnumerable<AgentStreamEvent> ResolvedRunStreamAsync(
-        PromptRequest request,
-        Action<AgentOutcome> setOutcome,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
-    {
-        AgentRunStream events =
-            (await ResolveAgentAsync()).RunStreamAsync(request, cancellationToken);
-
-        await foreach (AgentStreamEvent streamEvent in events.WithCancellation(cancellationToken))
-        {
-            yield return streamEvent;
-        }
-
-        setOutcome(events.Outcome);
-    }
-
-    // Composes once and reuses. Guarded because one agent serves prompts concurrently
-    // (SPEC.md §4.4) and an unguarded `??=` lets two arriving prompts each build a graph —
-    // two brokers over one audit sink, two of everything, one silently discarded.
-    //
-    // Registry selection is async (a registry can be a service call away) and a lock cannot
-    // hold an await, so resolution is three steps: reuse under the lock, select outside it,
-    // compose under it again. A builder call landing between the steps nulls the cache, so
-    // the very next resolution selects and composes afresh — nothing configured is lost.
-    private async ValueTask<IRunManagementService> ResolveAgentAsync()
-    {
-        IAgentRegistryBroker[] registries;
-
-        lock (this.compositionLock)
-        {
-            if (this.agent is not null)
-            {
-                return this.agent;
-            }
-
-            registries = [.. this.agentSources];
-        }
-
-        List<RegisteredAgent> registeredAgents = [];
-
-        foreach (IAgentRegistryBroker registry in registries)
-        {
-            registeredAgents.AddRange(await registry.SelectAgentsAsync());
-        }
-
-        lock (this.compositionLock)
-        {
-            return this.agent ??= Compose(registeredAgents);
-        }
-    }
-
-    // Every builder method drops the cached composition, so configuration set after a
-    // prompt still takes effect. Returning `this` without doing that would silently
-    // ignore the change.
-    /// <summary>
-    /// Puts the host's HTTP handler chain under every HTTP broker the agent composes: the brain,
-    /// the native brain, and MCP servers. The <b>External</b> seam for HTTP itself — pooled and
-    /// DNS-refreshing connections from <c>IHttpClientFactory</c>
-    /// (<c>IHttpMessageHandlerFactory.CreateHandler</c>), a proxy, a certificate, a resilience
-    /// handler, an observer — where a broker that built its own client could reach none of them
-    /// (principal review 2026-09-04, F-23).
-    /// </summary>
-    /// <remarks>
-    /// Ownership is explicit. A handler this delegate returns is yours: the broker wraps it in a
-    /// client that holds nothing of its own and never disposes it. Without this call every HTTP
-    /// broker creates its own handler, owned by the composition and released on the next one, so
-    /// a single long-lived agent holds one set for its life; short-lived agents composed by the
-    /// dozen should supply handlers from a factory instead. Order does not matter: brokers are
-    /// created at composition, so the handler reaches a brain or server registered before it.
-    /// </remarks>
-    /// <param name="handlers">Returns a handler for one broker; called once per HTTP broker.</param>
-    /// <returns>The same agent, so calls can be chained.</returns>
-    public StandardAgent Http(Func<HttpMessageHandler> handlers) =>
-        Set(() => this.httpHandlerSource = handlers);
-
-    private StandardAgent Set(Action configure)
-    {
-        lock (this.compositionLock)
-        {
-            configure();
-            this.agent = null;
-        }
-
-        return this;
-    }
-
-    // Reads an optional prompt file (constitution, consumption) resolved against the build
-    // output like skills. A missing file yields empty rather than an error, so a stale path
-    // degrades to the built-in guardian policy instead of bricking composition.
-    private static string ReadOptionalFile(IFileBroker file, string path)
-    {
-        if (string.IsNullOrWhiteSpace(path))
-        {
-            return string.Empty;
-        }
-
-        string fullPath = Path.Combine(AppContext.BaseDirectory, path);
-
-        return file.FileExists(fullPath)
-            ? file.ReadFile(fullPath)
-            : string.Empty;
-    }
-
-    // Assembles a guardian rubric from its parts in order, skipping any that are empty:
-    // constitution first (the law), then the policy (what to screen or score), then the
-    // contract (the output protocol the broker parses). The contract stays last so it is
-    // never displaced, whatever the policy above it.
-    private static string ComposeGuardianRubric(params string[] parts) =>
-        string.Join("\n\n", parts.Where(part => string.IsNullOrWhiteSpace(part) is false));
-
-    // One handler per HTTP broker: the host's when Http(...) was called, otherwise one this
-    // composition owns and the next composition releases.
-    private HttpMessageHandler CreateHttpHandler()
-    {
-        if (this.httpHandlerSource is not null)
-        {
-            return this.httpHandlerSource();
-        }
-
-        var handler = new HttpClientHandler();
-        this.ownedHttpHandlers.Add(handler);
-
-        return handler;
-    }
-
-    private void ReleaseOwnedHttpHandlers()
-    {
-        foreach (HttpMessageHandler handler in this.ownedHttpHandlers)
-        {
-            handler.Dispose();
-        }
-
-        this.ownedHttpHandlers.Clear();
-    }
-
-    private IRunManagementService Compose(IReadOnlyList<RegisteredAgent> registeredAgents)
-    {
-        ValidateComposition();
-        ReleaseOwnedHttpHandlers();
-
-        InferenceSettings? brain = this.brainSettings;
-
-        IFileBroker file = new FileBroker();
-
-        // The log is one broker; the decision log is another, applied over it by decoration
-        // when an audit sink is configured - over the built-in trace or a host's own logging
-        // broker alike, so .UseLogging(...) and .Audit(...) compose rather than compete (F-16).
-        ILoggingBroker trace =
-            this.loggingBroker ?? new LoggingBroker(
-                new NullLogger<LoggingBroker>(),
-                this.traceVerbosity,
-                string.IsNullOrEmpty(this.logPath) ? null : this.logPath);
-
-        IAuditBroker? audit =
-            this.auditBroker
-                ?? (string.IsNullOrEmpty(this.auditPath) ? null : new FileAuditBroker(this.auditPath));
-
-        // Composed here, above the log, because the same redaction that guards the model
-        // boundary guards the audit boundary (F-14); the generators and guardians below take it
-        // by decoration as before.
-        IRedactionBroker redaction = this.redactionBroker
-            ?? (this.redactionRules is null
-                ? new NotConfiguredRedactionBroker()
-                : new RuleRedactionBroker(this.redactionRules));
-
-        ILoggingBroker logging = audit is null
-            ? trace
-            : new AuditingLoggingBroker(
-                trace,
-                audit,
-                new TimeBroker(),
-                redaction,
-                new AuditPolicy(Payloads: this.auditPayloads),
-                () => this.identityResolver?.Invoke()?.Id);
-
-        // With only a native brain configured there is no V0 generator to build, and none is
-        // needed: SpeaksNatively routes every call to the V1 seam. The placeholder exists so the
-        // Brain foundation keeps one shape rather than growing a nullable dependency.
-        IGeneratorBroker generator =
-            this.generatorBroker
-                ?? (brain is null
-                    ? new FunctionGeneratorBroker((systemPrompt, userPrompt) =>
-                        throw new InvalidOperationException(
-                            "This agent has a native brain; the text protocol is not in use."))
-                    // The legacy-path values, resolved configured → framework default at
-                    // composition. A per-request value arrives on the request-carrying overload
-                    // and never through here.
-                    : new GeneratorBroker(
-                        CreateHttpHandler(),
-                        brain.ApiUrl, brain.ApiKey, brain.Model,
-                        brain.Temperature ?? ResolvedInference.DefaultTemperature,
-                        brain.MaxTokens ?? ResolvedInference.DefaultMaxTokens,
-                        brain.TimeoutSeconds));
-
-        // Without memory there is nothing to recall and no way to store: the foundation reads a
-        // not-configured broker, and the remember tool is never registered, so the model is not
-        // offered it (principal review 2026-09-04, F-05).
-        IMemoryService memoryService = this.memoryDisabled
-            ? new MemoryService(new NotConfiguredMemoryBroker(), logging)
-            : this.memoryBroker is null
-                ? new MemoryService(file, Path.GetFullPath(this.memoryPath), logging)
-                : new MemoryService(this.memoryBroker, logging);
-
-        List<ITool> allTools = this.memoryDisabled
-            ? [.. this.tools]
-            : [.. this.tools, new RememberTool(memoryService.RememberAsync)];
-
-        // The fleet materializes as tools — which is the whole design: a handoff is an act, so
-        // the advertisement opt-in, the perimeter, the audit and cancellation across the seam
-        // all apply because they already applied to tools. First to claim a name keeps it, the
-        // rule MCP servers live by, so a registry cannot shadow a tool the host wired in code.
-        HashSet<string> claimedNames =
-            new(allTools.Select(tool => tool.Name), StringComparer.OrdinalIgnoreCase);
-
-        foreach (RegisteredAgent registered in registeredAgents)
-        {
-            if (claimedNames.Add(registered.Name) is false)
-            {
-                continue;
-            }
-
-            allTools.Add(new AgentTool(
-                registered.Name,
-                registered.Agent,
-                AgentTool.GroundedHandoff,
-                registered.Description));
-        }
-
-        string constitution = ReadOptionalFile(file, this.constitutionPath);
-        string consumption = ReadOptionalFile(file, this.consumptionPath);
-
-        string gatePolicy = string.IsNullOrWhiteSpace(consumption)
-            ? GuardianPrompts.GatePolicy
-            : consumption;
-
-        string judgePolicy = string.IsNullOrWhiteSpace(consumption)
-            ? GuardianPrompts.JudgePolicy
-            : consumption;
-
-        string gateRubric = ComposeGuardianRubric(
-            constitution, gatePolicy, GuardianPrompts.GateContract);
-
-        string judgeRubric = ComposeGuardianRubric(
-            constitution, judgePolicy, GuardianPrompts.JudgeContract);
-
-        IClassifierBroker classifier =
-            this.classifierBroker
-                ?? (this.localGateScreen is not null
-                    ? new FunctionClassifierBroker(this.localGateScreen, gateRubric)
-                    : this.gateSettings is null
-                        ? new NotConfiguredClassifierBroker()
-                        : new ClassifierBroker(
-                            this.gateSettings.ApiUrl, this.gateSettings.ApiKey, this.gateSettings.Model,
-                            this.gateSettings.Temperature ?? 0.0, this.gateSettings.MaxTokens ?? 16,
-                            this.gateSettings.TimeoutSeconds, gateRubric));
-
-        IVerifierBroker verifier =
-            this.verifierBroker
-                ?? (this.localJudgeEvaluate is not null
-                    ? new FunctionVerifierBroker(this.localJudgeEvaluate, judgeRubric)
-                    : this.judgeSettings is null
-                        ? new NotConfiguredVerifierBroker()
-                        : new VerifierBroker(
-                            this.judgeSettings.ApiUrl, this.judgeSettings.ApiKey, this.judgeSettings.Model,
-                            this.judgeSettings.Temperature ?? 0.0, this.judgeSettings.MaxTokens ?? 16,
-                            this.judgeSettings.TimeoutSeconds, judgeRubric));
-
-        IToolBroker toolBroker = new ToolBroker(allTools);
-
-        // One source composes as itself; several compose behind the same seam the service
-        // already speaks — the tier above never learns how many integrations answered.
-        List<IMcpBroker> mcpBrokers = [.. this.mcpSources.Select(source => source())];
-
-        IMcpBroker mcp = mcpBrokers.Count switch
-        {
-            0 => new NotConfiguredMcpBroker(),
-            1 => mcpBrokers[0],
-            _ => new CompositeMcpBroker(mcpBrokers)
-        };
-
-        ISkillBroker skills = this.skillSources.Count switch
-        {
-            0 => new FileSkillBroker(Path.Combine(AppContext.BaseDirectory, "Skills")),
-            1 => this.skillSources[0],
-            _ => new CompositeSkillBroker(this.skillSources)
-        };
-
-        ISkillService skillService = new SkillService(skills, logging);
-
-        IKnowledgeService knowledgeService = this.knowledgeBroker is null
-            ? new KnowledgeService(
-                file,
-                Path.GetFullPath(this.knowledgePath),
-                this.knowledgePattern,
-                this.knowledgeMaxResults,
-                logging,
-                this.knowledgeMinScore)
-            : new KnowledgeService(this.knowledgeBroker, logging);
-
-        // One foundation over the remote tools, shared by the two natures that need it: Data
-        // advertises what the servers offer, Direction performs what the Brain chose. A single
-        // instance keeps discovery and execution reading the same servers.
-        IExternalToolService externalToolService = new ExternalToolService(mcp, logging);
-
-        // The Data nature, as two regions and the coordination that composes them. Retrieval is
-        // authored material selected by relevance; Recollection is what the agent accumulated.
-        DataCoordinationService data = new(
-            new RetrievalOrchestrationService(
-                skillService, knowledgeService, RenderToolCatalog(allTools), logging,
-                externalToolService,
-                RenderToolCatalogEntries(allTools)),
-            new RecollectionOrchestrationService(
-                memoryService,
-                new SessionService(
-                    this.sessionBroker ?? new NotConfiguredSessionBroker(), logging),
-                logging),
-            logging);
-
-        // One redaction broker across every model the agent drives. SPEC.md §4.6: the Gate
-        // screens the raw task and the Judge reads the task and the draft, and either may run on
-        // a different host than the Brain, so redacting only the Brain narrows nothing.
-        //
-        // It is applied by DECORATING each model broker rather than by handing it to each
-        // service. That is what makes "every model call" structural: a foundation holds one
-        // broker, knows nothing of redaction, and a fourth model call added tomorrow cannot
-        // forget (docs/architecture-alignment.md).
-        IResilienceBroker resilience =
-            this.resilienceBroker ?? new NotConfiguredResilienceBroker();
-
-        // Retry inside, redaction outside: the redacted prompt is what gets retried, so a value
-        // is tokenized once rather than once per attempt, and rehydration happens once at the end.
-        IGeneratorBroker generatorAtTheWire = new RedactingGeneratorBroker(
-            new RetryingGeneratorBroker(generator, resilience),
-            redaction);
-
-        IGeneratorBrokerV1? nativeBrain =
-            this.generatorBrokerV1 ?? this.nativeBrainSource?.Invoke();
-
-        IGeneratorBrokerV1? nativeAtTheWire = nativeBrain is null
-            ? null
-            : new RedactingGeneratorBrokerV1(
-                new RetryingGeneratorBrokerV1(nativeBrain, resilience),
-                redaction);
-
-        GateService gate = new(new RedactingClassifierBroker(classifier, redaction), logging);
-
-        // The Decision nature, as two regions and the judgment between them. Inference asks the
-        // model and reads its answer; Guardian screens what goes in and scores what comes out.
-        DecisionCoordinationService decision = new(
-            new InferenceOrchestrationService(
-                new BrainService(generatorAtTheWire, logging, nativeAtTheWire),
-                new UsageService(this.usageBroker, logging),
-                logging,
-                RenderToolDefinitions(allTools)),
-            new GuardianOrchestrationService(
-                gate,
-                new JudgeService(new RedactingVerifierBroker(verifier, redaction), logging),
-                new ContractService(this.contractBroker ?? new RuleContractBroker(), logging),
-                logging),
-            logging,
-            this.contractSchema);
-
-        // The allow-list is expressed as a policy, so the simple answer and an external policy
-        // engine travel one seam and a denial carries a reason either way (SPEC.md §4.9).
-        IPolicyBroker policy = this.policyBroker
-            ?? (this.allowedTools is null
-                ? new NotConfiguredPolicyBroker()
-                : new AllowListPolicyBroker(this.allowedTools));
-
-        IApprovalBroker approvals = this.approvalBroker
-            ?? (this.approvalRequiredTools is null
-                ? new NotConfiguredApprovalBroker()
-                : new RequireApprovalBroker(this.approvalRequiredTools));
-
-        // The Direction nature, as two regions and the order between them. Perimeter answers
-        // whether an act may happen; Execution performs it; the sequence belongs to neither.
-        DirectionCoordinationService direction = new(
-            new PerimeterOrchestrationService(
-                new PolicyService(policy, logging),
-                new ApprovalService(approvals, logging),
-                new EffectLedgerService(
-                    this.effectLedgerBroker ?? new InMemoryEffectLedgerBroker(),
-                    new TimeBroker(),
-                    logging),
-                new TimeBroker(),
-                logging),
-            new ExecutionOrchestrationService(
-                new InternalToolService(toolBroker, logging),
-                externalToolService,
-                new ReturnService(logging),
-                logging),
-            logging,
-            new PerimeterPolicy
-            {
-                Mode = this.permissionMode,
-                IrreversibleTools = [.. this.approvalRequiredTools ?? []],
-
-                DeclaredRisk = this.declaredRisk
-                    ?? new Dictionary<string, RiskLevel>(StringComparer.OrdinalIgnoreCase),
-
-                // What each tool says about itself, read once at composition. The tool is the
-                // only thing that knows what its arguments mean, and the framework never
-                // parses them.
-                ToolRisk = allTools.ToDictionary(
-                    tool => tool.Name,
-                    tool => tool.Risk,
-                    StringComparer.OrdinalIgnoreCase),
-
-                ToolScope = allTools.ToDictionary(
-                    tool => tool.Name,
-                    tool => (Func<string, string>)tool.ScopeOf,
-                    StringComparer.OrdinalIgnoreCase),
-
-                // Whether the allow-list speaks to an act at all — which the mode needs and a
-                // yes/no authorization decision cannot carry. Null when no allow-list was
-                // configured, so Ask asks about everything, which is what it says on the tin.
-                ExplicitlyPermits =
-                    policy is AllowListPolicyBroker allowList ? allowList.Mentions : null,
-
-                IdentityResolver = this.identityResolver,
-
-                EnforceSelection = this.enforceSelection,
-
-                // What selection could have offered, from the same rendering the catalogs use
-                // (SPEC.md §6.1: a description is the opt-in), so the perimeter can tell a
-                // withheld tool from an undescribed one and the two can never disagree.
-                AdvertisedTools = [.. Advertised(allTools).Select(tool => tool.Name)]
-            });
-
-        return new RunManagementService(
-            data, decision, direction, logging, this.maxTurns, new TimeBroker(), this.budget,
-            this.maxHistoryTurns, this.compensateOnFailure, this.screenToolOutput,
-            this.telemetryBroker, this.contractSchema, brain?.Temperature, brain?.MaxTokens,
-            allTools.Select(tool => tool.Name),
-            RenderToolNarrations(allTools),
-            this.localToolSelector is null ? null : new ToolSelector(this.localToolSelector),
-            Advertised(allTools).Select(tool => tool.Name),
-            this.identityResolver is null ? null : new PrincipalResolver(this.identityResolver));
-    }
-
-    // The catalog a "{{tools}}" marker in the agent's Data expands into. Only tools that
-    // carry a description are listed — a description is the opt-in (SPEC 6.1); a tool with
-    // none is callable but not advertised. Derived from the registered tools, so it never
-    // drifts from what is actually there.
-    // The same opt-in rule the text catalog uses (SPEC.md §6.1): a description is what offers a
-    // tool to the model. A tool without one stays callable but unadvertised, so declaring tools
-    // as data does not quietly widen what the Brain may reach for.
-    private static IReadOnlyList<ToolDefinition> RenderToolDefinitions(IEnumerable<ITool> tools) =>
-        [.. Advertised(tools)
-            .Select(tool => new ToolDefinition(tool.Name, tool.Description, tool.Parameters))];
-
-    // Narration templates, derived from the tools exactly as risk and scope are — the tool is
-    // what knows what its act means in the user's language, and derived data cannot drift.
-    private static IReadOnlyDictionary<string, ToolNarration> RenderToolNarrations(
-        IEnumerable<ITool> tools) =>
-        tools
-            .Where(tool =>
-                string.IsNullOrWhiteSpace(tool.NarrationStarting) is false
-                    || string.IsNullOrWhiteSpace(tool.NarrationObserved) is false)
-            .ToDictionary(
-                tool => tool.Name,
-                tool => new ToolNarration(tool.NarrationStarting, tool.NarrationObserved),
-                StringComparer.OrdinalIgnoreCase);
-
-    private static string RenderToolCatalog(IEnumerable<ITool> tools)
-    {
-        IEnumerable<string> describedTools = Advertised(tools)
-            .Select(tool => $"- {tool.Name} — {tool.Description} parameters: {tool.Parameters}");
-
-        return string.Join("\n", describedTools);
-    }
-
-    // The same catalog, per tool, so a run under selection (SPEC.md §4.15) can be shown only
-    // what it was offered. Derived from the same rendering as the whole-string catalog, so the
-    // two can never disagree about a tool's line.
-    private static IReadOnlyDictionary<string, string> RenderToolCatalogEntries(
-        IEnumerable<ITool> tools) =>
-        Advertised(tools).ToDictionary(
-            tool => tool.Name,
-            tool => $"- {tool.Name} — {tool.Description} parameters: {tool.Parameters}",
-            StringComparer.OrdinalIgnoreCase);
-
-    // What the model may be told about, in ONE place. The rule (SPEC.md §6.1: a description is
-    // the opt-in, and a tool without one stays callable but unadvertised) was written out twice —
-    // once for the text catalog Data renders into the prompt, once for the definitions Decision
-    // hands the native brain. Two copies of a rule about what the model is allowed to see is two
-    // chances to widen it by half, and a model told about a tool in the prompt but not in the
-    // schema — or the reverse — behaves differently on each protocol for no stated reason.
-    private static IEnumerable<ITool> Advertised(IEnumerable<ITool> tools) =>
-        tools.Where(tool => string.IsNullOrWhiteSpace(tool.Description) is false);
-
 }
